build: updated build system to allow build in-place
Sources can be rebuild and installed from already populated repositories on disk
which should speed-up developement cycle. Introduced fetch.sh which does force-checkout
of the source GIT repo.

Also revisited "clean" and "sample" targets to more automated ones that
dives into subprojects and call the "clean" /  "sample" target. This change
introduced warning of targets redefinition but this is intended because
some of the dependencies have extra parameters to the build.

Signed-off-by: Michal Jenikovsky <jendo@jmsystems.sk>